Vita: separate Vita-specific hacks into additional header
